Break out usePaginationEffects
